+++ src/api/types.ts
@@ -149,6 +149,10 @@ export interface ISupportCodeCoordinates {
    * Paths and/or glob expressions of user code to load, via `import()`
    */
   importPaths: string[]
+  /**
+   * Specifiers of loaders to register, via `register()`
+   */
+  loaders: string[]
 }
 
 /**
@@ -164,7 +168,7 @@ export interface ILoadSupportOptions {
    * feature file locations.
    */
   sources: ISourcesCoordinates
-  support: ISupportCodeCoordinates
+  support: Partial<ISupportCodeCoordinates>
 }
 
 /**
@@ -246,7 +250,7 @@ export interface IRunOptionsFormats {
  */
 export interface IRunConfiguration {
   sources: ISourcesCoordinates
-  support: ISupportCodeCoordinates
+  support: Partial<ISupportCodeCoordinates>
   runtime: IRunOptionsRuntime
   formats: IRunOptionsFormats
 }
@@ -277,7 +281,7 @@ export interface ISupportCodeLibrary {
  * mode, as parallel workers will each load the support code themselves anyway.
  */
 export type ISupportCodeCoordinatesOrLibrary =
-  | ISupportCodeCoordinates
+  | Partial<ISupportCodeCoordinates>
   | ISupportCodeLibrary
 
 /**